infrastructure/hpc_config.py
# initial infrastructure/hpc_config.py implementation
# class ColocationOptimizer:
#     def optimize_rack(self):
#         # Optimal server placement in NY4 data center
#         # Minimize fiber distance to NASDAQ matching engine
#         pass

import numa
from dpdk import *

class ColocationManager:
    def __init__(self):
        self.numa_nodes = numa.get_max_node() + 1
        self._bind_nic_queues()
        
    def _bind_nic_queues(self):
        # Bind NIC RX/TX queues to NUMA-local cores
        for q in range(16):
            core = q + numa.get_node_cpus(0)[0]
            Dpdk.bind_queue_to_core(q, core)
            
    def optimize_fpga_placement(self):
        """Place FPGAs on same NUMA node as trading logic"""
        fpga_numa = self._detect_fpga_topology()
        if fpga_numa != self.trading_thread_numa:
            self._migrate_trading_threads(fpga_numa)

# Deployment: AWS Snowblade instance
# Commands:
# hpc-optimize --rack NY4 --strategy ultra-low-latency