requirements.txt
yfinance==0.2.18
tensorflow==2.10.0
keras-tcn==3.4.0
cvxpy==1.3.1
talib-binary==0.4.24
curses-menu==0.6.0
matplotlib==3.7.1
pyfolio==0.9.2
numba==0.58.1
websockets==12.0
orjson==3.9.10
pymc==5.10.0
changefinder==0.3.1
bloomberg==1.4.3  # For institutional-grade data
dpdk==22.11
numactl==2.0.14
roce==1.2.1  # RDMA over Converged Ethernet
sec-edgar-api==3.0.2
cat-reporter==1.9.0
finra-compliance==0.8.11